RE-1544 Update gating hook dirs
We are moving from `gating/pre_merge_test` to `gating/check` and
`gating/post_merge_check` to `gating/periodic`. This commit updates
the gating hook dirs accordingly.

Additionally, we create symlinks for `gating/gate` and `gating/release`,
both of which point to `gating/check`. This can be adjusted by the
team at a later date.

Lastly, we create an empty `gating/generate_release_notes/run` script
as this will be executed on release and will fail the release if it
doesn't exist. The rpc-hummingbird team can eloborate on this in future
to ensure that release notes are generated automatically upon release.