RE-1617 Use rpc_component to download dependencies
This commit updates gating/check/run to use the component tool to
download registered dependencies specified in
component_requirements.yml. With this change in place, we can also
remove rpc-maas from ansible-role-test-requirements.yml.

Additionally, this commit removes run_tests.sh and symlinks it to
gating/check/run. The scripts are identical for the most part, so it
makes sense to remove the duplication by symlinking the manual testing
entry-point to what is getting executed in the gate.